优化docker-restart.sh逻辑，在solo进行部署时，只在镜像更新时重新部署镜像

+++ scripts/docker-restart.sh
@@ -7,13 +7,19 @@
 # 2. 可将该脚本加入 crontab，每日凌晨运行来实现自动更新
 #
 
-docker pull b3log/solo
-docker stop solo
-docker rm solo
-docker run --detach --name solo --network=host \
-    --env RUNTIME_DB="MYSQL" \
-    --env JDBC_USERNAME="root" \
-    --env JDBC_PASSWORD="123456" \
-    --env JDBC_DRIVER="com.mysql.cj.jdbc.Driver" \
-    --env JDBC_URL="jdbc:mysql://127.0.0.1:3306/solo?useUnicode=yes&characterEncoding=UTF-8&useSSL=false&serverTimezone=UTC" \
-    b3log/solo --listen_port=8080 --server_scheme=http --server_host=localhost
+isUpdate=$(docker pull b3log/solo|grep "Downloaded")
+
+if [[ -z  $isUpdate ]]
+then
+	echo This is the latest version
+else
+	docker stop solo
+	docker rm solo
+	docker run --detach --name solo --network=host \
+	--env RUNTIME_DB="MYSQL" \
+       	--env JDBC_USERNAME="root" \
+	--env JDBC_PASSWORD="123456" \
+	--env JDBC_DRIVER="com.mysql.cj.jdbc.Driver" \
+	--env JDBC_URL="jdbc:mysql://127.0.0.1:3306/solo?useUnicode=yes&characterEncoding=UTF-8&useSSL=false&serverTimezone=UTC" \
+       	b3log/solo --listen_port=8080 --server_scheme=http --server_host=localhost
+fi
